test(textarea): migrate card
