+++ core/src/components/textarea/test/card/textarea.e2e.ts-snapshots/textarea-card-md-ltr-Mobile-Firefox-linux.png
